Clean up starting logic to actually be functional
After the recent revelation that the cutscene variables don't work if gameinstructor is turned off, it became obvious that "new start logic" was the only *good* start logic, so the default start detection was removed, and now "New start logic" is the default behavior.

The code was also adjusted so that "autostart on cutsceneless" doesn't completely clobber the functionality of the timer if new start logic was also enabled (given that this will now be the default behavior). It seems there was an attempt to make them coexist peacefully in the past, but that no longer worked - if you tried to reset the timer while you still had control/weren't watching a cutscene, the timer would just continuously start, which means it wouldn't even work at all to literally autostart on a cutsceneless campaign. Anyways, things should hopefully be less confusing now for the end user, for however much longer this autosplitter is used until SST takes over.

Lastly, I updated the name of the autostart on cutsceneless setting to hopefully be slightly clearer, and added a little bit of extra info to the tooltip.