Sigscanning hasControl
I have no idea how reliable this is but it works on 1.0.0.0, 1.0.0.5, 1.0.2.2, 1.0.3.3, 1.0.4.1 (newest), and 1.0.0.5 on my second PC. It also works on L4D2 2.0.0.0 and 2.0.9.1! xd

No idea if more stuff should be wildcarded, but I'm just shocked that I was able to figure any of this out in the first place. ADHD hyperfocus + cheat engine = giga stonks