Backporting features to L4D1 autosplitter
We now match the version to existing offsets in the case of 1.0.0.5 and 1.0.4.1, and the timer won't start if cheats is on. Also, adjusted the failsafe delay to 125ms because it gets annoying at 250 in l4d1 (resetting timer during loading screen too late causing timer not to start when it should otherwise be perfectly fine to do so). Timer start should probably just be overhauled at some point maybe