Simplify/improve start logic
Removed the setting that enables cutsceneless start, as the autosplitter can now seamlessly use both start methods (cutsceneless and cutscene) without the user needing to specify. Hooray.

I also upped the failsafe bit on the cutscene thing to 250ms, just to be safe (not that it would realistically matter anyway, but it was changed to 125ms for the wrong reasons, and it didn't actually help solve the problem it was intended to solve).

Overall, still far from perfect, but it's Better Than Before.