Hopefully, maybe, but probably not fix the splitting on intro cutscene bug
Currently there's a bug best explained by Twitch clips:

https://clips.twitch.tv/ObedientAwkwardPeachCharlietheUnicorn-Paj1gG0NOMkceBqp
https://clips.twitch.tv/SmokySmellyHawkChocolateRain-Mm5KWpT_xEl21VvE

This might not fix those egregious instances of the bug, however it will (hopefully) prevent the cases where it does randomly split a single time. I would test this more to see if it actually fixes it, but it literally took me 10 minutes to trigger the single-split bug to draw this conclusion on how I might fix it, so I quite frankly don't have the patience to sit there and roll for it.